Support better type checking of array sections
Previously it would just return the type of the array, not taking into
account what the section is referring to, which is somewhat ok for 1
dimensional arrays but for more dimensions it really should tell us the
expression is a single dimensional array.

This helps us check whether interfaces usage is valid, which does care
about the dimensionality.